Made environment variable optional, added error checks.
DPHOME:
If the environment variables DPHOME and/or DP_3RDPARTY_PATH are not set,
defaults based on the source root directory are used for compilation.
The runtime home directory needs to include a new file dp.cfg.
Searched for this is $DPHOME, the executable directory and the
compile time home (set via CMAKE).
This way compilation and running the sames works without the need
to set any environment variables but it still can be set optionally.

TerrainRendering:
* heightmap, texturemap and resolution parameters are mandatory, so
a check was added to quit with an error instead of a crash.
* The offset was set to the resolution parameter instead of the offset
parameter.
* Check if resolution parameter has 3 values added.
* Terminate after printing help - if you need help, other parameters are
likely to be wrong or missing anyway.

MaterialExchanger:
* The check for a wrong number of command line parameters now also
terminates the program.

Added more verbose error messages in case files or Plugins were not found.

Throwing and catching exceptions instead of error messages.

Change-Id: Iace79465c474a9b97b52b8f24666660f3cb12800
Reviewed-on: http://git-master/r/1130992
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Andreas Suessenbach <asuessenbach@nvidia.com>
Tested-by: Andreas Suessenbach <asuessenbach@nvidia.com>